server/modal.com/modal-cpu/requirements.txt
fastapi==0.112.2
pydantic==2.8.2
llama-cpp-python==0.3.5
faiss-cpu==1.8.0
numpy==1.24.0
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
sentence-transformers==2.2.2
huggingface-hub==0.22.2
transformers==4.38.2
--extra-index-url https://download.pytorch.org/whl/cpu